feat(vae): complete #45 — seam-free tiled decode + bf16 default (fixes fp16 NaN)
Finishes the remaining/at-risk pieces of the VAE-decode work (#45), validated
against the real intrusive-memory/sdxl-vae-fp16-mlx checkpoint:

- Tiled decode: replace the hard-crop stitch with a feathered weighted
  overlap-add (the diffusers `enable_vae_tiling` / ComfyUI strategy), accumulated
  in float32. The global mid-block attention still drifts tile-to-tile, but the
  feather blend spreads that drift across the overlap band instead of leaving a
  seam. Real-weights validation: tiled-vs-full PSNR 30.5 dB, tile-boundary /
  interior gradient ratio 1.20 (a seam would spike this). Still opt-in via
  `decodeTileLatentSize`; single-pass default behavior is unchanged.

- Decode dtype: change the default from fp16 to **bfloat16**. The real SDXL VAE
  is fp16-unstable (`force_upcast=true`) — an fp16 decode overflows to NaN
  (proven by a new regression guard). bf16 has fp16's 16-bit footprint (so the
  activation-memory win is kept) with fp32-class range. Real-weights parity vs
  fp32: PSNR 59.0 dB, NaN-free. This fixes a latent output-corruption bug in the
  prior fp16 default that synthetic CPU weights did not surface.

- Tests: new `SDXLVAEDecoderRealWeightsTests` GPU suite loading the real
  checkpoint (skip-on-absent; stage to /tmp for sandboxed xctest) — tiled seam
  parity, bf16-vs-fp32 parity, and the fp16-overflow guard. New `featherWeights`
  unit test; relax the synthetic multi-tile assertion to a tolerance (the blend
  is a normalized weighted average, not a hard crop). Update the default-dtype
  config test to bf16.

All TuberiaCatalog (161) + TuberiaCatalogGPU (55) tests pass on arm64.

Closes #45. Unblocks intrusive-memory/Vinetas#92 (macOS 4K).

+++ Sources/TuberiaCatalog/Decoders/SDXLVAEDecoderConfiguration.swift
@@ -10,35 +10,40 @@ public struct SDXLVAEDecoderConfiguration: Sendable {
   public let latentChannels: Int
   /// VAE latent scaling factor (applied internally by the decoder).
   public let scalingFactor: Float
-  /// Compute dtype for the VAE decode pass. Defaults to fp16 to halve the
-  /// decode-time activation footprint (#45). The VAE weights ship as fp16 and
-  /// the output is clipped to 8-bit downstream, so a float32 decode wastes ~2×
-  /// the activation memory for precision that is discarded anyway. Set to
-  /// `.float32` to restore the original full-precision decode if a quality
-  /// regression is observed, or `.bfloat16` if an intermediate overflows fp16.
+  /// Compute dtype for the VAE decode pass. Defaults to **bfloat16** to halve the
+  /// decode-time activation footprint vs float32 (#45) while staying numerically
+  /// safe: the SDXL VAE is fp16-unstable (`force_upcast=true` in its config) and
+  /// an fp16 decode of the real checkpoint overflows to NaN (validated on real
+  /// weights in `SDXLVAEDecoderRealWeightsTests`). bfloat16 has the same 16-bit
+  /// footprint as fp16 but fp32-class dynamic range, so it captures the memory
+  /// win without the overflow. Set to `.float32` to restore the original
+  /// full-precision decode; `.float16` is **not** safe for this checkpoint.
   public let decodeDType: DType
   /// Tile edge in *latent* pixels for tiled decode (#45). When non-nil and the
   /// latent is larger than the tile in either spatial dimension, the decode is
   /// performed tile-by-tile to bound the activation transient independently of
-  /// output resolution. `nil` (the default) preserves the original
-  /// single-pass, behavior-identical decode.
+  /// output resolution. `nil` (the default) preserves the original single-pass,
+  /// behavior-identical decode — callers opt in for high-resolution output where
+  /// the single-pass transient would be too large (e.g. 4K).
   ///
-  /// - Important: Tiled decode is currently **experimental and off by default**.
-  ///   The SDXL VAE mid-block self-attention is global over the bottleneck, so
-  ///   each tile attends only over its own spatial extent — small tiles can
-  ///   shift low-frequency tone tile-to-tile and leave visible seams. See
-  ///   `SDXLVAEDecoderModel.decodeTiled`.
+  /// Tiles are recombined with a feathered weighted overlap-add (see
+  /// `SDXLVAEDecoderModel.decodeTiled`), which keeps the tile boundaries
+  /// seam-free even though the VAE mid-block attention is global — validated on
+  /// real weights in `SDXLVAEDecoderRealWeightsTests` (tiled-vs-full PSNR > 30 dB,
+  /// boundary-gradient ratio ≈ 1.2). It is left off by default because the
+  /// single-pass decode is marginally higher-fidelity and cheaper at resolutions
+  /// that already fit comfortably in memory.
   public let decodeTileLatentSize: Int?
-  /// Halo (in *latent* pixels) added on each side of a tile before decoding and
-  /// cropped back off after, used to suppress seam artifacts in tiled decode.
-  /// Ignored when `decodeTileLatentSize` is nil.
+  /// Halo (in *latent* pixels) added on each side of a tile before decoding; the
+  /// feather ramp spans this halo so adjacent tiles blend smoothly. Must be > 0
+  /// for the blend to work. Ignored when `decodeTileLatentSize` is nil.
   public let decodeTileLatentOverlap: Int
 
   public init(
     componentId: String = "sdxl-vae-decoder-fp16",
     latentChannels: Int = 4,
     scalingFactor: Float = 0.13025,
-    decodeDType: DType = .float16,
+    decodeDType: DType = .bfloat16,
     decodeTileLatentSize: Int? = nil,
     decodeTileLatentOverlap: Int = 8
   ) {
